+++ src/XBullet.EasyTesting.Snapshots/Properties/AssemblyInfo.cs
@@ -1,0 +1,3 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("XBullet.EasyTesting.Tests")]