+++ src/XBullet.EasyTesting/Properties/AssemblyInfo.cs
@@ -2,3 +2,4 @@ using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("XBullet.EasyTesting.Azure")]
 [assembly: InternalsVisibleTo("XBullet.EasyTesting.Http")]
+[assembly: InternalsVisibleTo("TestApi.IntegrationTests")]